DEV-358: Split Service Requests and Patient screens
