+++ hospital_simulator/app/(DashboardLayout)/components/service-requests/service-request-table.tsx
@@ -37,7 +37,6 @@ const ServiceRequestTable: React.FC<Props> = ({ rows }) => {
     const columns: GridColDef[] = [
         { field: 'lastUpdated', headerName: 'Last Updated', type: 'dateTime', flex: 2 },
         { field: 'title', headerName: 'Title', flex: 3 },
-        { field: 'patient', headerName: 'Patient', flex: 2 },
         { field: 'reasonReference', headerName: 'Reason', flex: 2 },
         {
             field: 'status',
