+++ hospital_simulator/app/(DashboardLayout)/layout/sidebar/MenuItems.tsx
@@ -1,14 +1,5 @@
 import {
-  IconAperture,
-  IconCheckupList,
-  IconCopy,
   IconEyeCog,
-  IconFlame,
-  IconLayoutDashboard,
-  IconLogin,
-  IconMoodHappy,
-  IconTypography,
-  IconUserPlus,
 } from "@tabler/icons-react";
 
 import { uniqueId } from "lodash";
@@ -16,40 +7,13 @@ import { uniqueId } from "lodash";
 const Menuitems = [
   {
     navlabel: true,
-    subheader: "Home",
-  },
-
-  {
-    id: uniqueId(),
-    title: "Dashboard",
-    icon: IconLayoutDashboard,
-    href: "/",
-  },
-  {
-    navlabel: true,
-    subheader: "FHIR",
+    subheader: "EHR",
   },
   {
     id: uniqueId(),
-    title: "Service Requests",
-    icon: IconCheckupList,
-    href: "/service-requests",
-  },
-  {
-    navlabel: true,
-    subheader: "Auth",
-  },
-  {
-    id: uniqueId(),
-    title: "Login",
-    icon: IconLogin,
-    href: "/authentication/login",
-  },
-  {
-    id: uniqueId(),
-    title: "Register",
-    icon: IconUserPlus,
-    href: "/authentication/register",
+    title: "Patients",
+    icon: IconEyeCog,
+    href: "/patients",
   },
 ];
