+++ hospital_simulator/app/(DashboardLayout)/page.tsx
@@ -1,41 +1,14 @@
 'use client'
-import { Grid, Box } from '@mui/material';
-import PageContainer from '@/app/(DashboardLayout)/components/container/PageContainer';
-// components
-import ProfitExpenses from '@/app/(DashboardLayout)/components/dashboard/ProfitExpenses';
-import TrafficDistribution from '@/app/(DashboardLayout)/components/dashboard/TrafficDistribution';
-import UpcomingSchedules from '@/app/(DashboardLayout)/components/dashboard/UpcomingSchedules';
-import TopPayingClients from '@/app/(DashboardLayout)/components/dashboard/TopPayingClients';
-import ProductSales from '@/app/(DashboardLayout)/components/dashboard/ProductSales';
+import {useRouter} from "next/navigation";
+import {useEffect} from "react";
 
 const Dashboard = () => {
-  return (
-    <PageContainer title="Dashboard" description="this is Dashboard">
-      <Box>
-        <Grid container spacing={3}>
-          <Grid size={{xs:12, lg:8}}>
-            <ProfitExpenses />
-          </Grid>
-          <Grid size={{xs:12, lg:4}}>
-            <Grid container spacing={3}>
-              <Grid size={{xs:12}}>
-                <TrafficDistribution />
-              </Grid>
-              <Grid size={{xs:12}}>
-                <ProductSales />
-              </Grid>
-            </Grid>
-          </Grid>
-          <Grid size={{xs:12, lg:4}}>
-            <UpcomingSchedules />
-          </Grid>
-          <Grid size={{xs:12, lg:8}}>
-            <TopPayingClients />
-          </Grid>
-        </Grid>
-      </Box>
-    </PageContainer>
-  )
+    const { push } = useRouter();
+
+    useEffect(() => {
+        push('/patients');
+    }, []);
+    return <p></p>;
 }
 
 export default Dashboard;
